typing Comp par2
