+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@types/react-redux": "^7.1.20",
     "antd": "^4.6.2",
     "node-sass": "^4.14.1",
     "normalize.css": "^8.0.1",
