+++ src/Types/Types.ts
@@ -2,7 +2,11 @@ export enum SortOfTickets {
   cheapest = "cheapest",
   fastest = "fastest",
 }
-enum FiltersStateType {}
+export enum TranslateType {
+  cheapest = "Самый дешевый",
+  fastest = "Самый быстрый",
+}
+enum FiltersStateType { }
 
 export interface InitialFiltersStateType {
   [key: string]: boolean;
